Fix tests/compilation for JS/native. Introduce platform-specific SttpClientExceptionExtensions
